Updated MeshTests for skeleton bind pose changes (#15)
* Fix for TestSendEvents

* Modify mesh tests for bind pose removal

Mesh tests now use SXRSkin.setInverseBindPose instead of SXRSkeleton.setBindPose

* 1. put VIdeoTests and WidgetTests in their own sections
2. remove VertexColorShader, PhongShader now supports vertex colors

* Moved testSendEvents to io category

* Wait a frame after centering model in tests

* add waits to allow tests more time before screenshot

* add waits for asset tests, add missing golden masters

* put back x3d tests

* add waits in asset import tests

* fix texture count in test

* add delays in in external scene test

* fixed compile error in TestSendEvents

* scale model based on skeleton

* replaced golden master for some lighting tests

assimp fixed a bug and now several lighting tests look lots better

* fix centering in animation tests

* Change centering code and golden masters

Changes to skinning import now preserves the original position of the skeleton's bones, previously we were using the bind pose. Some golden masters have to change because objects are positioned differently.

* Change centering code and golden masters

Changes to skinning import now preserves the original position of the skeleton's bones, previously we were using the bind pose. Some golden masters have to change because objects are positioned differently.

* added GLTF skinning test

* disable join vertices in skeleton test

testSkeletonAnimation will crash in the join vertices process in assimp. The Collada loader produces morph targets with a different vertex count than the base mesh when importing Andromeda.dae (not sure why). This causes it to crash if the JoinVertices post-process step is enabled

* remove gray background in skeleton test

* fix TREX skinning golden master

* new golden masters for embedded texture test

newer assimp has different lighting for FBX import, grill and license plate are brighter

* new golden master for testRemoveLight

Now astroboy is in the animated position because of animation import changes

* move light in point light test

The golden master we have is for an unlit scene. The original is black which is correct because the light is behind the object. This PR moves the light in front of the object and puts in a new golden master.

* use correct golden master

* updated position in skeleton bind pose test

* bounding volume tests wait for texture load